Implemented an improved backup system which will back up the database as a zip file to be extracted. Maximum backups can be set via discordbot.yaml backup_quantity. It will delete the oldest log once the number of backups exceeds that number.
I also implemented this deletion system for logs, its set to 15 max logs for now and is not configurable. I figured it would be enough for 15 start, stop bot sessions and the user can manually back them up if they would like. I'll probably change this in the future to be configurable.